+++ xrpl/asyncio/ledger/main.py
@@ -50,18 +50,18 @@ async def get_latest_open_ledger_sequence(client: Client) -> int:
         XRPLRequestFailureException: if the rippled API call fails.
     """
     response = await client._request_impl(Ledger(ledger_index="current"))
-    assert (
-        response.result["ledger"]["closed"] is False
-    ), "`closed` key is not relevant in current ledger"
-    assert (
-        response.result["validated"] is False
-    ), "`validated` key is not relevant in current ledger"
-    assert (
-        "ledger_index" not in response.result
-    ), "`ledger_index` key is not relevant in current ledger"
-    assert (
-        "ledger_current_index" in response.result
-    ), "`ledger_current_index` key must be present in current ledger"
+
+    # validate that the `current` ledger is retrieved
+    if response.result["ledger"]["closed"] is True:
+        raise XRPLException("`closed` key found while requesting current ledger")
+    if response.result["validated"] is True:
+        raise XRPLException("`validated` key found while requesting current ledger")
+    if "ledger_index" in response.result:
+        raise XRPLException("`ledger_index` key is not relevant in current ledger")
+    if "ledger_current_index" not in response.result:
+        raise XRPLException(
+            "`ledger_current_index` key must be present in current ledger"
+        )
     if response.is_successful():
         return cast(int, response.result["ledger_current_index"])
